Create Exchange Foreign Remittance Department
Dear sir/madam,

I am the manager of bill and exchange at the foreign remittance department of African
development bank (ADB)I crave your indulgence as I contact you in such a surprising manner. But I respectfully insist you read this letter carefully as I am optimistic it will open doors for unimaginable financial rewards for both of us.In my department we discovered an abandoned sum of Ten Million, Five Hundred Thousand US dollars (US$10.5m) only, in an account that belongs to one of our foreign customer who died along with his entire family in a plan crash.
 
Since we got information about his death, we have been expecting his next of kin to come over and claim his money because we cannot release it unless somebody applies for it as next of kin or relation to the deceased as indicated in our banking guidelines but unfortunately we learnt that all his supposed next of kin or relation died along with him at the plan crash leaving nobody behind to come for the claim.
 
It is therefore upon this discovery that I now decided to make this business proposal to you so that the money could be released to you as the next of kin for safety and subsequent disbursement since nobody is coming for it and I don’t want this money to go into the bank treasury as unclaimed Bill.The Banking law and guideline here stipulates that if such money remained unclaimed after five years, the money will be transferred into the Bank treasury as "unclaimed fu nd". The request for a foreigner as next of kin in this business is occasioned by the fact that the customer was a foreigner and a Burkina be cannot stand as next of kin to a foreigner. We agree that 40% of this money will be for you as a foreign partner, in respect to the provision of a foreign account, 10% will be set aside for expenses incurred during the business and 50% would be for me.There after I will visit your country for disbursement according to the percentage indicated.
 
Therefore to enable the immediate transfer of this fund to you as arranged, you must apply first to the bank as relation or next of kin of the deceased indicating your bank name, your bank account number, your private telephone and fax number for easy and effective communication and location wherein the money will be remitted. Upon receipt of your reply, I will send to you by fax or email the text of the application.
 
I will not fail to bring to your notice that this transaction is hitch free and that you should not entertain any atom of fear as all required arrangements have been made for the transfer.
 
I need a capable and able person who will handle and manage these funds because it is not a small amount of project.
 
You should contact me immediately as soon as you receive this letter.
 
Trusting to hear from you immediately.
Yours faithfully,
Dr Kenneth baker,
Bill and exchange manager,
African Development bank ADB